+++ Almanac/NPC/DialogueManager.cs
@@ -11,7 +11,6 @@ using Almanac.TreasureHunt;
 using Almanac.UI;
 using Almanac.Utilities;
 using BepInEx;
-using HarmonyLib;
 using PieceManager;
 using ServerSync;
 using UnityEngine;